+++ blockly/assets/levels/dfs_1.level
@@ -1,4 +1,4 @@
-ICE
+DEFAULT
 2,23
 
 WWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWW
