+++ blockly/assets/levels/leveltemplate_1.level
@@ -1,4 +1,4 @@
-FOREST
+DEFAULT
 2,23
 
 WWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWWW
