+++ blockly/src/starter/PathfinderStarter.java
@@ -13,13 +13,13 @@ import core.utils.components.path.SimpleIPath;
 import java.io.IOException;
 import java.util.logging.Level;
 import level.AiMazeLevel;
-import systems.LevelEditorSystem;
+import contrib.systems.LevelEditorSystem;
 import systems.PathfindingSystem;
 import utils.CheckPatternPainter;
 import utils.pathfinding.DFSPathFinding;
 
 /** This class starts the dungeon Ai level to visualize the DFS and BFS. */
-public class KiStarter {
+public class PathfinderStarter {
   private static final boolean DRAW_CHECKER_PATTERN = true;
 
   /**
